+++ packages/api/src/middlewares/LogMiddleware.ts
@@ -8,6 +8,13 @@ const LOGGER_SECRET_FIELDS = ["password", "token"];
 
 const LOGGER_IGNORED_ROUTES = [/^\/docs/, /^\/assets/];
 
+function recursiveFilter(obj: object) {
+    Object.entries(obj).forEach(([key, value]) => {
+        if (LOGGER_SECRET_FIELDS.includes(key)) obj[key] = "**********";
+        else if (typeof value === "object" && value) recursiveFilter(obj[key]);
+    });
+}
+
 export const expressLogger = () =>
     expressWinston.logger({
         transports: [
@@ -39,14 +46,14 @@ export const expressLogger = () =>
             if (LOGGER_IGNORED_ROUTES.some(regex => regex.test(req.url))) return true;
             return false;
         },
-        requestFilter: (req, propName) => (LOGGER_SECRET_FIELDS.includes(propName) ? "**********" : req[propName]),
+        requestFilter: (req, propName) => {
+            if (propName === "body" && typeof req[propName] === "object" && req[propName])
+                recursiveFilter(req[propName]);
+            return LOGGER_SECRET_FIELDS.includes(propName) ? "**********" : req[propName];
+        },
         responseFilter: (req, propName) => {
-            if (propName === "body" && req[propName]) {
-                LOGGER_SECRET_FIELDS.forEach(field => {
-                    if (!req[propName][field]) return;
-                    req[propName][field] = "**********";
-                });
-            }
+            if (propName === "body" && typeof req[propName] === "object" && req[propName])
+                recursiveFilter(req[propName]);
             return LOGGER_SECRET_FIELDS.includes(propName) ? "**********" : req[propName];
         },
     });