+++ packages/api/src/modules/notify/outPipes/BrevoContactNotifyPipe.ts
@@ -138,6 +138,7 @@ export class BrevoContactNotifyPipe extends BrevoNotifyPipe implements NotifyOut
             [AgentJobTypeEnum.ADMINISTRATOR]: "Gestionnaire administratif et financier",
             [AgentJobTypeEnum.EXPERT]: "Chargé de mission / Expert métier",
             [AgentJobTypeEnum.SERVICE_HEAD]: "Responsable de service",
+            [AgentJobTypeEnum.CONTROLLER]: "Contrôleur / Inspecteur",
             [AgentJobTypeEnum.OTHER]: "Autre",
         };
 
@@ -170,12 +171,9 @@ export class BrevoContactNotifyPipe extends BrevoNotifyPipe implements NotifyOut
             service: "SERVICE",
             phoneNumber: "TELEPHONE",
             jobType: "BUYER_PERSONNAE",
-            centralStructure: "ADMIN_CENTRALE",
+            structure: "STRUCTURE",
             decentralizedLevel: "COMPETENCE_TERRITORIALE",
             decentralizedTerritory: "ECHELON_TERRITORIAL",
-            decentralizedStructure: "SERVICE_DECONCENTRE",
-            operatorStructure: "OPERATEUR",
-            territorialStructure: "COLLECTIVITES",
             territorialScope: "ECHELON_COLLECTIVITE",
             lastname: "NOM",
             firstname: "PRENOM",
@@ -205,10 +203,9 @@ export class BrevoContactNotifyPipe extends BrevoNotifyPipe implements NotifyOut
 
         const updateContact = new Brevo.UpdateContact();
         const attributes: Record<string, string | boolean> = buildAttributesObject(data);
-        attributes.COMPTE_ACTIVE = true;
+        attributes.COMPTE_ACTIVE = true; // TODO le rendre configurable ?
         updateContact.attributes = attributes;
         updateContact.listIds = SENDIND_BLUE_CONTACT_LISTS;
-        console.log({ email: data.email, updateContact });
         return this.apiInstance
             .updateContact(data.email, updateContact)
             .then(() => true)