+++ packages/front/src/lib/components/StructureFormStep/DecentralizedSubStep/DecentralizedSubStep.svelte
@@ -23,7 +23,7 @@
 
 <Radio
     options={ctrl.levelOptions}
-    label="Vous êtes :"
+    label="Sélectionnez votre périmètre :"
     bind:value={values.decentralizedLevel}
     on:change={({ detail }) => ctrl.onChoosingLevel(detail)} />
 
@@ -42,7 +42,7 @@
             <AutocompleteSelect
                 options={$regionOptions}
                 bind:value={values.decentralizedTerritory}
-                label="Quelle est votre région"
+                label="Quelle est votre région ?"
                 on:change
                 placeholder="Ex : Occitanie" />
         </div>
@@ -55,7 +55,7 @@
                 bind:value={values.structure}
                 label="Quelle est votre administration ?"
                 on:change
-                placeholder="Ex : DETS59, Préfecture" />
+                placeholder="Ex : DDETS59, Préfecture" />
         {:else}
             <Input
                 id="structure"
@@ -63,7 +63,7 @@
                 bind:value={values.structure}
                 label="Quelle est votre administration ?"
                 on:change
-                placeholder="Ex : DETS59, Préfecture" />
+                placeholder="Ex : DDETS59, Préfecture" />
         {/if}
     </div>
 </fieldset>